// src/stores/storage/interfaces.ts
export interface Storage {
    amount: string;
    amountHashed: string;
    email: string;
    emailHashed: string;
    phone: string;
    phoneHashed: string;
    firstName: string;
    firstNameHashed: string;
    lastName: string;
    lastNameHashed: string;
    city: string;
    cityHashed: string;
    state: string;
    stateHashed: string;
};